improve search for VTK-9
Improve the find_package calls with vtk-9 cmake config files.
In this version, the variable VTK_MODULES_ENABLED has been
replaced by VTK_AVAILABLE_COMPONENTS.

The additional change of the components names, which have their
'vtk' prefix removed in vtk-9 is not covered by this patch for
compatibility reasons.

See the forum thread noted below for a description of this:
https://forum.freecadweb.org/viewtopic.php?f=4&t=60068

Signed-off-by: Bernd Waibel <waebbl-gentoo@posteo.net>