Sketcher: fix external edge tolerance issue
The tolerance problem occurs when using plane normal angles to
determin if a circle should be projected into an ellipse, because the
tolerance of angle may not reflect the tolerance of projected length.

This commit checks the difference in projected and original radius to
decide whether to import the external edge as a circle or ellipse.

This commit also adds support for projected elliptic arc.

Fixes realthunder/FreeCAD_Assembly3#933